+++ engines/cengine/include/avdl_localisation.h
@@ -37,6 +37,8 @@ void avdl_localisation_print(struct avdl_localisation *o);
 
 void avdl_locale_set(const char *locale_code);
 
+const char *avdl_locale_getSystemLocale(const char *locale_code);
+
 #ifdef __cplusplus
 }
 #endif