+++ engines/cengine/src/avdl_localisation_cpp.cpp
@@ -1,0 +1,10 @@
+#include "avdl_localisation.h"
+
+#if defined( AVDL_STEAM )
+const char * avdl_locale_getSystemLocale(const char *locale_code) {
+	if (strcmp(SteamApps()->GetCurrentGameLanguage(), "greek") == 0) {
+		return "el";
+	}
+	return "en";
+}
+#endif