+++ metadata/changelog.md
@@ -1,1 +1,2 @@
 * Temporarily allow up to 400 characters for localisation values
+* On Steam, automaticall choose default language